+++ datarobot_provider/example_dags/model_compute_predictions_dag.py
@@ -6,42 +6,26 @@
 #
 # Released under the terms of DataRobot Tool and Utility Agreement.
 from datetime import datetime
-import datarobot as dr
 
 from airflow.decorators import dag
 from airflow.decorators import task
+from datarobot import PredictJob
 
 from datarobot_provider.hooks.datarobot import DataRobotHook
-from datarobot_provider.operators.model_insights import ComputeFeatureImpactOperator
 from datarobot_provider.operators.model_predictions import AddExternalDatasetOperator
 from datarobot_provider.operators.model_predictions import RequestModelPredictionsOperator
-from datarobot_provider.operators.prediction_explanations import (
-    PredictionExplanationsInitializationOperator,
-    ComputePredictionExplanationsOperator,
-)
 from datarobot_provider.sensors.model_insights import DataRobotJobSensor
-
-"""
-Example DAG to create prediction explanations for the specified dataset.
-In order to create Prediction Explanations for a particular model and dataset, you must use the next steps:
- - Compute feature impact for the model using ComputeFeatureImpactOperator
- - Initialize prediction explanations for the model using PredictionExplanationsInitializationOperator
- - Compute predictions for the model and dataset using RequestModelPredictionsOperator
- - Compute predictions explanations for the model and dataset using ComputePredictionExplanationsOperator
-"""
 
 
 @dag(
     schedule=None,
     start_date=datetime(2023, 1, 1),
-    tags=['example', 'dataset', 'model'],
-    # Default json config example:
-    params={'threshold_high': 0.9, 'threshold_low': 0.1, 'max_explanations': 3},
+    tags=['example', 'dataset'],
 )
 def compute_model_predictions(
-    project_id='64a7f21f1110f0df910ddb4f',
-    model_id="64a7f27e8e0fd5cae6282a8b",
-    dataset_id="64bfc36171d728b6fa2e369c",
+    project_id=None,
+    model_id=None,
+    dataset_id=None,
 ):
     if not project_id:
         raise ValueError("Invalid or missing `project_id` value")
@@ -54,20 +38,6 @@ def compute_model_predictions(
         task_id="add_external_dataset",
         project_id=project_id,
         dataset_id=dataset_id,
-    )
-
-    compute_feature_impact_op = ComputeFeatureImpactOperator(
-        task_id="compute_feature_impact",
-        project_id=project_id,
-        model_id=model_id,
-    )
-
-    feature_impact_complete_sensor = DataRobotJobSensor(
-        task_id="feature_impact_complete",
-        project_id=project_id,
-        job_id=compute_feature_impact_op.output,
-        poke_interval=5,
-        timeout=3600,
     )
 
     request_model_predictions_op = RequestModelPredictionsOperator(
@@ -85,76 +55,27 @@ def compute_model_predictions(
         timeout=3600,
     )
 
-    # In order to compute prediction explanations you have to initialize it for a particular model.
-    prediction_explanations_initialization_op = PredictionExplanationsInitializationOperator(
-        task_id="prediction_explanations_initialization",
-        project_id=project_id,
-        model_id=model_id,
-    )
-
-    prediction_explanations_initialization_sensor = DataRobotJobSensor(
-        task_id="prediction_explanations_initialization_complete",
-        project_id=project_id,
-        job_id=prediction_explanations_initialization_op.output,
-        poke_interval=5,
-        timeout=3600,
-    )
-
-    compute_prediction_explanations_op = ComputePredictionExplanationsOperator(
-        task_id="compute_prediction_explanations",
-        project_id=project_id,
-        model_id=model_id,
-        external_dataset_id=add_external_dataset_op.output,
-    )
-
-    compute_prediction_explanations_sensor = DataRobotJobSensor(
-        task_id="compute_prediction_explanations_complete",
-        project_id=project_id,
-        job_id=compute_prediction_explanations_op.output,
-        poke_interval=5,
-        timeout=3600,
-    )
-
     @task(task_id="example_custom_python_code")
-    def using_custom_python_code(
-        pe_project_id, pe_model_id, predict_job_id, datarobot_conn_id='datarobot_default'
-    ):
+    def using_custom_python_code(project_id, predict_job_id, datarobot_conn_id='datarobot_default'):
         """Example of using custom python code:"""
 
         # Initialize DataRobot client
         DataRobotHook(datarobot_conn_id).run()
 
         # Should be used only with a very small datasets (not higher than a few megabytes).
-        predictions_df = dr.PredictJob.get_predictions(
-            project_id=pe_project_id, predict_job_id=predict_job_id
-        )
-        # Get last prediction explanations computation
-        explanations_id = dr.PredictionExplanations.list(pe_project_id, model_id=pe_model_id)[-1].id
-        prediction_explanations_df = dr.PredictionExplanations.get(
-            pe_project_id, explanations_id
-        ).get_all_as_dataframe()
+        output_df = PredictJob.get_predictions(project_id=project_id, predict_job_id=predict_job_id)
 
-        # Put your logic with the model predictions output and prediction explanations here, for example:
-        return predictions_df["positive_probability"].mean(), prediction_explanations_df[
-            'explanation_0_feature'
-        ].describe().get('top')
+        # Put your logic with the model predictions output dataframe here, for example:
+        return output_df["positive_probability"].mean() > 0.5
 
     example_custom_python_code = using_custom_python_code(
-        pe_project_id=project_id,
-        pe_model_id=model_id,
-        predict_job_id=request_model_predictions_op.output,
+        project_id=project_id, predict_job_id=request_model_predictions_op.output
     )
 
     (
         add_external_dataset_op
-        >> compute_feature_impact_op
-        >> feature_impact_complete_sensor
         >> request_model_predictions_op
         >> model_predictions_sensor
-        >> prediction_explanations_initialization_op
-        >> prediction_explanations_initialization_sensor
-        >> compute_prediction_explanations_op
-        >> compute_prediction_explanations_sensor
         >> example_custom_python_code
     )
